[LaraApi] Extend set operations to the class itself (prototype)
In previous commits, the set operations were only available as static
methods that create new sets. This commit supports the same operations
on the `prototype` chain, hence can be applied to an exisitng StringSet
object and mutate it. The previous static methods were refactored to
take advantage of the new `prototype` versions