Joinpoint methods that handle enums now use string types
Affects the creation of abstract joinpoints from the Language specification.
If a method returns or receives a enum type also defined in the Language Specification then the type is considered to be a String. This is because all enum members are forcibly strings and this allows for proper interface with JS APIs.